+++ lib/avromatic/model/messaging_serialization.rb
@@ -6,14 +6,15 @@ module Avromatic
     # This concern adds support for serialization based on AvroTurf::Messaging.
     # This serialization leverages a schema registry to prefix encoded values
     # with an id for the schema.
-    module Messaging
+    module MessagingSerialization
       extend ActiveSupport::Concern
 
-      delegate :messaging, to: :Avromatic
+      delegate :avro_messaging, to: :class
+      private :avro_messaging
 
       module Encode
         def avro_message_value
-          messaging.encode(
+          avro_messaging.encode(
             value_attributes_for_avro,
             schema_name: value_avro_schema.fullname
           )
@@ -21,7 +22,7 @@ module Avromatic
 
         def avro_message_key
           raise 'Model has no key schema' unless key_avro_schema
-          messaging.encode(
+          avro_messaging.encode(
             key_attributes_for_avro,
             schema_name: key_avro_schema.fullname
           )
@@ -29,17 +30,19 @@ module Avromatic
       end
       include Encode
 
-      # This module provides methods to message_decode an Avro-encoded value and
+      # This module provides methods to decode an Avro-encoded value and
       # an optional Avro-encoded key as a new model instance.
       module Decode
 
         # If two arguments are specified then the first is interpreted as the
         # message key and the second is the message value. If there is only one
         # arg then it is used as the message value.
-        def message_decode(*args)
+        def avro_message_decode(*args)
           message_key, message_value = args.size > 1 ? args : [nil, args.first]
-          key_attributes = message_key && messaging.decode(message_key, schema_name: key_avro_schema.fullname)
-          value_attributes = messaging.decode(message_value, schema_name: avro_schema.fullname)
+          key_attributes = message_key &&
+            avro_messaging.decode(message_key, schema_name: key_avro_schema.fullname)
+          value_attributes = avro_messaging
+            .decode(message_value, schema_name: avro_schema.fullname)
 
           new(value_attributes.merge!(key_attributes || {}))
         end
@@ -48,7 +51,9 @@ module Avromatic
       module ClassMethods
         # The messaging object acts as an intermediary talking to the schema
         # registry and using returned/specified schemas to decode/encode.
-        delegate :messaging, to: :Avromatic
+        def avro_messaging
+          Avromatic.messaging
+        end
 
         include Decode
       end