+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## v0.5.0
 - Rename `Avromatic::Model::Decoder` to `MessageDecoder`.
-- Rename `.deserialize` on generated models to `.message_decode`.
-- Add `#avro_raw_key`, `#avro_raw_value` and `.raw_decode` methods to
+- Rename `.deserialize` on generated models to `.avro_message_decode`.
+- Add `#avro_raw_key`, `#avro_raw_value` and `.avro_raw_decode` methods to
   generated models to support encoding and decoding without a schema registry.
 
 ## v0.4.0